chore: move init function to the different package
